Refactor to use a word search dialog container
