+++ djAerolith/wordwalls/static/js/wordwalls/newtable/search/dialog.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import SearchRows from './search_rows';
+import SearchRows from './rows';
 
 const WordSearchDialog = props => (
   <div className="row" style={{ marginTop: '8px' }}>
@@ -18,6 +18,7 @@ const WordSearchDialog = props => (
         style={{ marginTop: '0.75em' }}
         onClick={props.onSearchSubmit}
         data-dismiss="modal"
+        disabled={props.disabled ? 'disabled' : ''}
       >Play!
       </button>
       <button
@@ -25,6 +26,7 @@ const WordSearchDialog = props => (
         style={{ marginTop: '0.75em', marginLeft: '1em' }}
         onClick={props.onFlashcardSubmit}
         data-dismiss="modal"
+        disabled={props.disabled ? 'disabled' : ''}
       >Flashcard
       </button>
     </div>
@@ -47,6 +49,7 @@ WordSearchDialog.propTypes = {
   removeSearchRow: PropTypes.func.isRequired,
   onSearchSubmit: PropTypes.func.isRequired,
   onFlashcardSubmit: PropTypes.func.isRequired,
+  disabled: PropTypes.bool.isRequired,
 };
 
 export default WordSearchDialog;
