+++ package.json
@@ -28,6 +28,7 @@
     "babel-preset-react": "^6.24.1",
     "compression-webpack-plugin": "^1.0.1",
     "css-loader": "^0.25.0",
+    "enzyme": "^3.3.0",
     "eslint": "^4.8.0",
     "eslint-config-airbnb": "^16.0.0",
     "eslint-plugin-compat": "^2.0.1",
@@ -44,6 +45,7 @@
     "less-loader": "^2.2.3",
     "react-test-renderer": "^16.1.1",
     "rimraf": "^2.6.2",
+    "sinon": "^6.1.3",
     "style-loader": "^0.13.1",
     "text-loader": "^0.0.1",
     "url-loader": "^0.5.7",
@@ -70,6 +72,7 @@
     "mustache": "^2.3.0",
     "promise-polyfill": "^7.0.0",
     "prop-types": "^15.6.0",
+    "qs": "^6.5.2",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-dropzone": "^4.1.3",
